requirements.txt
asgiref==3.8.1
Django==5.2.1
django-bootstrap5==25.1
django-decouple==2.1
django-phonenumber-field[phonenumbers]
sqlparse==0.5.3
tzdata==2025.2
